renamed progress to progressCalculation to differentiate it from userdata.mjs
