+++ casual.html
@@ -1520,7 +1520,7 @@ Taking the Iron Key off the Goblin in the first rat area and making a save will 
 <script type="module">
 	import exportNamespace from './js/module-shim.mjs';
 	import * as common from './js/cellGenerator.mjs';
-	import * as progress from './js/progress.mjs';
+	import * as progress from './js/progressCalculation.mjs';
 	import * as userdata from './js/userdata.mjs';
 	exportNamespace(common);
 	exportNamespace(progress);
