+++ index.html
@@ -95,7 +95,7 @@
 	import exportNamespace from './js/module-shim.mjs'
 	import * as obliviondata from './js/obliviondata.mjs'
 	import * as common from './js/cellGenerator.mjs';
-	import * as progress from './js/progress.mjs'
+	import * as progress from './js/progressCalculation.mjs'
 	import * as userdata from './js/userdata.mjs'
 	import * as sharing from './js/sharing.mjs'
 	window.common = common;
