+++ js/progressCalculation.mjs
