+++ speedrun-3.html
@@ -4744,7 +4744,7 @@ You have completed the 100% run!
 <script type="module">
 	import exportNamespace from './js/module-shim.mjs'
 	import * as common from './js/cellGenerator.mjs';
-	import * as progress from './js/progress.mjs'
+	import * as progress from './js/progressCalculation.mjs'
 	import * as userdata from './js/userdata.mjs'
 	import * as obliviondata from './js/obliviondata.mjs'
 	import * as sharing from './js/sharing.mjs';
